LayoutTests/js/dom/string-prototype-scopes-expected.txt

This test checks whether String.prototype methods correctly handle |this| if |this| becomes window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS String(error) is "TypeError: Type error"
PASS successfullyParsed is true

TEST COMPLETE
